Move estimateHistorySize function to internal_utils.go
