+++ internal/internal_event_handlers.go
@@ -942,7 +942,7 @@ func (weh *workflowExecutionEventHandlerImpl) ProcessEvent(
 		return err
 	}
 
-	historySum := weh.estimateHistorySize(event)
+	historySum := estimateHistorySize(weh.logger, event)
 	atomic.AddInt64(&weh.workflowInfo.TotalHistoryBytes, int64(historySum))
 
 	// When replaying histories to get stack trace or current state the last event might be not
@@ -1383,140 +1383,4 @@ func (weh *workflowExecutionEventHandlerImpl) handleSignalExternalWorkflowExecut
 	signal.handle(nil, err)
 
 	return nil
-}
-
-func (weh *workflowExecutionEventHandlerImpl) estimateHistorySize(event *m.HistoryEvent) int {
-	sum := historySizeEstimationBuffer
-	switch event.GetEventType() {
-	case m.EventTypeWorkflowExecutionStarted:
-		if event.WorkflowExecutionStartedEventAttributes != nil {
-			sum += len(event.WorkflowExecutionStartedEventAttributes.Input)
-			sum += len(event.WorkflowExecutionStartedEventAttributes.ContinuedFailureDetails)
-			sum += len(event.WorkflowExecutionStartedEventAttributes.LastCompletionResult)
-			sum += sizeOf(event.WorkflowExecutionStartedEventAttributes.Memo.GetFields())
-			sum += sizeOf(event.WorkflowExecutionStartedEventAttributes.Header.GetFields())
-			sum += sizeOf(event.WorkflowExecutionStartedEventAttributes.SearchAttributes.GetIndexedFields())
-		}
-	case m.EventTypeWorkflowExecutionCompleted:
-		if event.WorkflowExecutionCompletedEventAttributes != nil {
-			sum += len(event.WorkflowExecutionCompletedEventAttributes.Result)
-		}
-	case m.EventTypeWorkflowExecutionSignaled:
-		if event.WorkflowExecutionSignaledEventAttributes != nil {
-			sum += len(event.WorkflowExecutionSignaledEventAttributes.Input)
-		}
-	case m.EventTypeWorkflowExecutionFailed:
-		if event.WorkflowExecutionFailedEventAttributes != nil {
-			sum += len(event.WorkflowExecutionFailedEventAttributes.Details)
-		}
-	case m.EventTypeDecisionTaskStarted:
-		if event.DecisionTaskStartedEventAttributes != nil {
-			sum += getLengthOfStringPointer(event.DecisionTaskStartedEventAttributes.Identity)
-		}
-	case m.EventTypeDecisionTaskCompleted:
-		if event.DecisionTaskCompletedEventAttributes != nil {
-			sum += len(event.DecisionTaskCompletedEventAttributes.ExecutionContext)
-			sum += getLengthOfStringPointer(event.DecisionTaskCompletedEventAttributes.Identity)
-			sum += getLengthOfStringPointer(event.DecisionTaskCompletedEventAttributes.BinaryChecksum)
-		}
-	case m.EventTypeDecisionTaskFailed:
-		if event.DecisionTaskFailedEventAttributes != nil {
-			sum += len(event.DecisionTaskFailedEventAttributes.Details)
-		}
-	case m.EventTypeActivityTaskScheduled:
-		if event.ActivityTaskScheduledEventAttributes != nil {
-			sum += len(event.ActivityTaskScheduledEventAttributes.Input)
-			sum += sizeOf(event.ActivityTaskScheduledEventAttributes.Header.GetFields())
-		}
-	case m.EventTypeActivityTaskStarted:
-		if event.ActivityTaskStartedEventAttributes != nil {
-			sum += len(event.ActivityTaskStartedEventAttributes.LastFailureDetails)
-		}
-	case m.EventTypeActivityTaskCompleted:
-		if event.ActivityTaskCompletedEventAttributes != nil {
-			sum += len(event.ActivityTaskCompletedEventAttributes.Result)
-			sum += getLengthOfStringPointer(event.ActivityTaskCompletedEventAttributes.Identity)
-		}
-	case m.EventTypeActivityTaskFailed:
-		if event.ActivityTaskFailedEventAttributes != nil {
-			sum += len(event.ActivityTaskFailedEventAttributes.Details)
-		}
-	case m.EventTypeActivityTaskTimedOut:
-		if event.ActivityTaskTimedOutEventAttributes != nil {
-			sum += len(event.ActivityTaskTimedOutEventAttributes.Details)
-			sum += len(event.ActivityTaskTimedOutEventAttributes.LastFailureDetails)
-		}
-	case m.EventTypeActivityTaskCanceled:
-		if event.ActivityTaskCanceledEventAttributes != nil {
-			sum += len(event.ActivityTaskCanceledEventAttributes.Details)
-		}
-	case m.EventTypeMarkerRecorded:
-		if event.MarkerRecordedEventAttributes != nil {
-			sum += len(event.MarkerRecordedEventAttributes.Details)
-		}
-	case m.EventTypeWorkflowExecutionTerminated:
-		if event.WorkflowExecutionTerminatedEventAttributes != nil {
-			sum += len(event.WorkflowExecutionTerminatedEventAttributes.Details)
-		}
-	case m.EventTypeWorkflowExecutionCanceled:
-		if event.WorkflowExecutionCanceledEventAttributes != nil {
-			sum += len(event.WorkflowExecutionCanceledEventAttributes.Details)
-		}
-	case m.EventTypeWorkflowExecutionContinuedAsNew:
-		if event.WorkflowExecutionContinuedAsNewEventAttributes != nil {
-			sum += len(event.WorkflowExecutionContinuedAsNewEventAttributes.Input)
-			sum += len(event.WorkflowExecutionContinuedAsNewEventAttributes.FailureDetails)
-			sum += len(event.WorkflowExecutionContinuedAsNewEventAttributes.LastCompletionResult)
-			sum += sizeOf(event.WorkflowExecutionContinuedAsNewEventAttributes.Memo.GetFields())
-			sum += sizeOf(event.WorkflowExecutionContinuedAsNewEventAttributes.Header.GetFields())
-			sum += sizeOf(event.WorkflowExecutionContinuedAsNewEventAttributes.SearchAttributes.GetIndexedFields())
-		}
-	case m.EventTypeStartChildWorkflowExecutionInitiated:
-		if event.StartChildWorkflowExecutionInitiatedEventAttributes != nil {
-			sum += len(event.StartChildWorkflowExecutionInitiatedEventAttributes.Input)
-			sum += len(event.StartChildWorkflowExecutionInitiatedEventAttributes.Control)
-			sum += sizeOf(event.StartChildWorkflowExecutionInitiatedEventAttributes.Memo.GetFields())
-			sum += sizeOf(event.StartChildWorkflowExecutionInitiatedEventAttributes.Header.GetFields())
-			sum += sizeOf(event.StartChildWorkflowExecutionInitiatedEventAttributes.SearchAttributes.GetIndexedFields())
-		}
-	case m.EventTypeChildWorkflowExecutionCompleted:
-		if event.ChildWorkflowExecutionCompletedEventAttributes != nil {
-			sum += len(event.ChildWorkflowExecutionCompletedEventAttributes.Result)
-		}
-	case m.EventTypeChildWorkflowExecutionFailed:
-		if event.ChildWorkflowExecutionFailedEventAttributes != nil {
-			sum += len(event.ChildWorkflowExecutionFailedEventAttributes.Details)
-			sum += getLengthOfStringPointer(event.ChildWorkflowExecutionFailedEventAttributes.Reason)
-		}
-	case m.EventTypeChildWorkflowExecutionCanceled:
-		if event.ChildWorkflowExecutionCanceledEventAttributes != nil {
-			sum += len(event.ChildWorkflowExecutionCanceledEventAttributes.Details)
-		}
-	case m.EventTypeSignalExternalWorkflowExecutionInitiated:
-		if event.SignalExternalWorkflowExecutionInitiatedEventAttributes != nil {
-			sum += len(event.SignalExternalWorkflowExecutionInitiatedEventAttributes.Control)
-			sum += len(event.SignalExternalWorkflowExecutionInitiatedEventAttributes.Input)
-		}
-
-	default:
-		weh.logger.Warn("unknown event type", zap.String("Event Type", event.GetEventType().String()))
-	}
-
-	return sum
-}
-
-// simple function to estimate the size of a map[string][]byte
-func sizeOf(o map[string][]byte) int {
-	sum := 0
-	for k, v := range o {
-		sum += len(k) + len(v)
-	}
-	return sum
-}
-
-func getLengthOfStringPointer(s *string) int {
-	if s == nil {
-		return 0
-	}
-	return len(*s)
 }
